+++ tge-extra/ui.c
@@ -12,3 +12,12 @@ void tge_draw_modal(TGE_Canvas *canvas, const char *title,
     tge_draw_centered_text(canvas, height / 2 + 1, subtitle, TGE_COLOR_WHITE,
                            TGE_COLOR_BLACK);
 }
+
+void tge_draw_region(TGE_Canvas *canvas, TGE_Rect rect, const char *title,
+                     TGE_Color fg)
+{
+    tge_draw_frame(canvas, rect.x, rect.y, rect.w, rect.h, fg,
+                   TGE_COLOR_DEFAULT);
+    if (title)
+        tge_draw_text(canvas, rect.x + 1, rect.y, title, fg, TGE_COLOR_DEFAULT);
+}